refactor encryption to just scrypt
