+++ packages/jellyfish-wallet-encrypted/__tests__/provider/scrypt_simple.test.ts
@@ -1,4 +1,4 @@
-import { SimpleScryptsy } from '../../src/scrypt/scrypt_simple'
+import { SimpleScryptsy } from '../../src'
 
 describe('passphraseToKey()', () => {
   it('Should be able to generate from random (same) passphrase into multiple desired length secret', () => {
