+++ packages/jellyfish-wallet-encrypted/__tests__/scrypt.test.ts
@@ -1,21 +1,20 @@
-import { SimpleScryptsy, SymmetricEncryption } from '../src'
+import { Scrypt, SimpleScryptsy } from '../src'
 
-const simpleScryptsy = new SimpleScryptsy()
-const encryption = new SymmetricEncryption(simpleScryptsy)
+const scrypt = new Scrypt(new SimpleScryptsy())
 
 it('should be able to encrypt / decrypt', async () => {
   const privKey = 'e9873d79c6d87dc0fb6a5778633389f4e93213303da61f20bd67fc233aa33262'
   const passphrase = 'password'
   const buffer = Buffer.from(privKey, 'hex')
 
-  const data = encryption.encrypt(buffer, passphrase)
+  const data = scrypt.encrypt(buffer, passphrase)
   expect(data).not.toStrictEqual(null)
   expect(data.encode()).not.toStrictEqual(null)
 
   const encoded = data.encode()
   const hash = data.hash.toString('hex')
 
-  const decrypted = await encryption.decrypt(encoded, passphrase, hash)
+  const decrypted = await scrypt.decrypt(encoded, passphrase, hash)
   expect(decrypted).toStrictEqual(buffer)
   expect(decrypted.toString('hex')).toStrictEqual(privKey)
 })
@@ -25,14 +24,14 @@ it('should be able to encrypt / decrypt - simple passphrase, a 6 digit pin', asy
   const passphrase = '135790'
   const buffer = Buffer.from(privKey, 'hex')
 
-  const data = encryption.encrypt(buffer, passphrase)
+  const data = scrypt.encrypt(buffer, passphrase)
   expect(data).not.toStrictEqual(null)
   expect(data.encode()).not.toStrictEqual(null)
 
   const encoded = data.encode()
   const hash = data.hash.toString('hex')
 
-  const decrypted = await encryption.decrypt(encoded, passphrase, hash)
+  const decrypted = await scrypt.decrypt(encoded, passphrase, hash)
   expect(decrypted).toStrictEqual(buffer)
   expect(decrypted.toString('hex')).toStrictEqual(privKey)
 })
@@ -42,14 +41,14 @@ it('Should work with variable data length - long', async () => {
   const passphrase = 'passcode'
   const buffer = Buffer.from(privKey, 'hex')
 
-  const data = encryption.encrypt(buffer, passphrase)
+  const data = scrypt.encrypt(buffer, passphrase)
   expect(data).not.toStrictEqual(null)
   expect(data.encode()).not.toStrictEqual(null)
 
   const encoded = data.encode()
   const hash = data.hash.toString('hex')
 
-  const decrypted = await encryption.decrypt(encoded, passphrase, hash)
+  const decrypted = await scrypt.decrypt(encoded, passphrase, hash)
   expect(decrypted).toStrictEqual(buffer)
   expect(decrypted.toString('hex')).toStrictEqual(privKey)
 })
@@ -59,14 +58,14 @@ it('should work with variable data length - short', async () => {
   const passphrase = 'passcode'
   const buffer = Buffer.from(privKey, 'hex')
 
-  const data = encryption.encrypt(buffer, passphrase)
+  const data = scrypt.encrypt(buffer, passphrase)
   expect(data).not.toStrictEqual(null)
   expect(data.encode()).not.toStrictEqual(null)
 
   const encoded = data.encode()
   const hash = data.hash.toString('hex')
 
-  const decrypted = await encryption.decrypt(encoded, passphrase, hash)
+  const decrypted = await scrypt.decrypt(encoded, passphrase, hash)
   expect(decrypted).toStrictEqual(buffer)
   expect(decrypted.toString('hex')).toStrictEqual(privKey)
 })
@@ -74,7 +73,7 @@ it('should work with variable data length - short', async () => {
 it('should accept "even" number data length', async () => {
   const buffer = Buffer.from('eeffaabb', 'hex')
 
-  const data = await encryption.encrypt(buffer, 'password')
+  const data = await scrypt.encrypt(buffer, 'password')
   expect(data).not.toStrictEqual(null)
 })
 
@@ -82,11 +81,11 @@ it('should reject "odd" number data length', async () => {
   const buffer = Buffer.from('eeffaa', 'hex')
 
   await expect(async () => {
-    await encryption.encrypt(buffer, 'password')
+    await scrypt.encrypt(buffer, 'password')
   }).rejects.toThrow('Data length must be even number')
 })
 
 it('should pad "odd" number data length', async () => {
   const padded = Buffer.from('00' + 'eeffaa', 'hex')
-  encryption.encrypt(padded, 'password')
+  scrypt.encrypt(padded, 'password')
 })
