+++ packages/jellyfish-wallet-encrypted/src/provider/scrypt_provider.ts
