+++ packages/jellyfish-wallet-encrypted/src/provider/scrypt_simple.ts
